Updated addon to include the elusive 4th pet on each team; also added support for its auras. Updated addon to know if we can even make a move on our turn and if we can't to not inform the program. Added enums (static) to more easily identify what a value is when coding. Added a PetAction class to track if an ability is still active when using AI. Removed direct access to QLists from QML replaced with more invokables and properties. QML testing almost complete.

+++ WoWPetBattler/Move.cpp
@@ -1,7 +1,7 @@
 #include "Move.h"
 
 //Constructor
-Move::Move(int move)
+Move::Move(quint8 move)
 	: move(move)
 {
 	this->heuristic = 0;
